requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
langchain
langchain-community
langchain-text-splitters
faiss-cpu
rank-bm25
sentence-transformers
python-dotenv
pydantic